refactor(calculator): extract TasksSection into its own component
The largest section of CalculatorView (~225 lines of JSX covering
flat-rate rows, device/action comboboxes, modifier chips, and the
shared qty/price/total/delete strip) moves out, taking TaskTotalsRow
and the tagSuggestions helper with it. The parent shrinks by ~320 lines
and the unused Combobox import is gone.
